ecommerce_platform/requirements.txt
agate==1.9.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
comm==0.2.3
confluent-kafka==2.3.0
cryptography==46.0.5
daff==1.4.2
dbt-adapters==1.9.0
dbt-common==1.12.0
dbt-core==1.8.7
dbt-duckdb==1.8.0
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.5.1
debugpy==1.8.20
decorator==5.2.1
deepdiff==7.0.1
defusedxml==0.7.1
duckdb==1.1.0
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
Faker==24.0.0
fastapi==0.110.0
fastjsonschema==2.21.2
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.46
great-expectations==0.18.12
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
importlib-metadata==6.11.0
ipykernel==6.31.0
ipython==8.18.1
ipywidgets==8.1.8
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
leather==0.4.1
Logbook==1.5.3
makefun==1.16.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mashumaro==3.20
matplotlib-inline==0.2.1
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.18.1
nbclient==0.10.2
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
ordered-set==4.1.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.6
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.4.0
plotly==6.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.27.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.15
scipy==1.13.1
Send2Trash==2.1.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.36.3
streamlit==1.32.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomli==2.4.0
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.27.0
wcwidth==0.6.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
zipp==3.23.0
